archive-service/requirements.txt
matrix-nio>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pydantic>=2.5.0
pydantic-settings>=2.1.0
aiohttp>=3.9.0
